Fix problem setting the querydict object
On line 522 of this module, a previous edit added the "remove_quotes" function around the val variable. I reverted this code to the way it was done before support for $eq was added - the val variable in many cases is a list which causes the remove_quotes function to fail.  When it is a list, the quotes have already been removed